requirements.txt
streamlit==0.78.0
Pillow==8.1.2

numpy~=1.20.1
matplotlib~=3.3.4